+++ src/modules/normalizer.cpp
@@ -10,7 +10,7 @@ Tensor EmpiricalNormalizer::forward(const Tensor& observations) {
 void EmpiricalNormalizer::update(const Tensor& observations) {
   this->count_ += observations.size(0);
 
-  const double rate = static_cast<double>(observations.size(0)) / this->count_;
+  const float rate = static_cast<float>(observations.size(0)) / this->count_;
   const Tensor& var_obs =
       torch::var(observations, /*dim=*/0, /*unbiased=*/true);
   const Tensor& mean_obs = torch::mean(observations, /*dim=*/0);
